Add ReadWriteTimeout property to ServiceClientBase
I added the TimeSpan? ReadWriteTimeout property to ServiceClientBase class to support two different timeout values as HttpWebRequest does.
(The other one is the Timeout property that is already implemented)

The HttpWebRequest.Timeout is the timeout to reach the HTTP server and the ReadWriteTimeout is the additional timeout value to get the response data.

By having this two different timeouts, you can (for example) set the .Timeout to 5000 and .ReadWriteTimeout to a much larger timeout value if you want to download a lot of data and if the url is wrong, or the server is down, you don't have to wait till the long timeout happens.