Eliminate default member values from route query
Generating a route from a request DTO currently produces redundant
parameters for value types that are equal to their default value. This
change modifies RouteMember so that it only returns a non-null value to
be contributed to the route if the current value is not equal to that
type's default value.