Config to force Swagger to use https for all URLs
In load balanced environments, it's common to terminate SSL at the load
balancer. This makes requests come in with a scheme of http, even though
the client is actually calling them with https. ServiceStack sees the
http and incorrectly formats URLs that are sent back to the client.

When EndpointHostConfig.ForceHttps is enabled, outgoing URLs should
always have a scheme of https, regardless of the scheme on incoming
requests.

At the moment, this is only implemented for Swagger.