Failing test for Html.CheckBoxFor() - cannot submit bool=true
Html.CheckBoxFor creates a second hidden input for the bool to ensure that
the false value is set with the post. But in the case of a true value,
both the checkbox and hidden input values are sent with the form.
ServiceStack sees this data as a list and has trouble deserializing the
list into a bool.