Updated Silverlight to the latest code - 3.9.50.
Not superfluous, the previous version was broken. One user even reported it,
but we mistakenly did not test the released version but the (at the time)
current version of the source. For reference, the broken version - still on
NuGet - is 3.9.44; don't let this version fool you, SL5 build lag behind the
official build number so 3.9.49 includes SL5 binaries 3.9.44.
The crash happened in ServiceClientBase's constructor due to the misuse of
the CookieContainer.
Current code works fine, let's update SL5 build and update NuGet...